requirements.txt
Django==3.1.5
django-solo==1.1.5
django-imagekit==4.0.2
Pillow==8.1.0
